Fixing drop table migrations

+++ db/migrate/20170902233427_drop_tables_not_used.rb
@@ -1,6 +1,6 @@
 class DropTablesNotUsed < ActiveRecord::Migration[5.1]
   def change
-      drop_table :santiagos
-      drop_table :examples
+      drop_table :santiagos if ActiveRecord::Base.connection.table_exists?
+      drop_table :examples if ActiveRecord::Base.connection.table_exists?
   end
 end
